refactor(global): move the failed-install cleanup into its own module
globalAdd and globalUpdate mock globalActivation.js wholesale for
activateGlobalInstall, which used to take cleanupFailedGlobalInstall down
with it and left both suites reimplementing the cleanup contract in their
mocks. Pulling the real one back in through jest.requireActual is not an
option either: requiring an ES module needs the synchronous vm APIs Node
only exposes from 24.9, so the suites failed to run on Node 22 and 24.

Give the cleanup and the error-message helpers it shares with activation
their own modules. The command suites now mock only what they intend to and
run the shipped cleanup, so the fresh-install removal and both-errors
assertions cover production code on every supported Node version.
