+++ pnpm11/global/commands/test/globalUpdate.test.ts
@@ -20,9 +20,6 @@ const readInstalledPackages = jest.fn<(installDir: string) => Promise<Array<{ al
 const summaryDebug = jest.fn()
 const activateGlobalInstall = jest.fn<(opts: unknown) => Promise<Set<string>>>().mockResolvedValue(new Set(['fresh']))
 const cleanupReplacedGlobalInstalls = jest.fn<(opts: unknown) => Promise<void>>().mockResolvedValue(undefined)
-// The real cleanup runs here: these tests cover how the command wires it up,
-// while its diagnostics are asserted in globalActivation.test.ts.
-const { cleanupFailedGlobalInstall } = jest.requireActual<typeof import('../src/globalActivation.js')>('../src/globalActivation.js')
 
 jest.unstable_mockModule('@pnpm/core-loggers', () => ({ summaryLogger: { debug: summaryDebug } }))
 jest.unstable_mockModule('@pnpm/global.packages', () => ({
@@ -36,7 +33,6 @@ jest.unstable_mockModule('@pnpm/global.packages', () => ({
 jest.unstable_mockModule('../src/checkGlobalBinConflicts.js', () => ({ checkGlobalBinConflicts }))
 jest.unstable_mockModule('../src/globalActivation.js', () => ({
   activateGlobalInstall,
-  cleanupFailedGlobalInstall,
   cleanupReplacedGlobalInstalls,
 }))
 jest.unstable_mockModule('../src/installGlobalPackages.js', () => ({ installGlobalPackages }))
